+++ src/components/aotc-2023/in-game-overlay.tsx
@@ -4,7 +4,7 @@ import * as styles from './in-game-overlay.css';
 export function AOTC2023InGameOverlay({ children }: React.PropsWithChildren<{}>) {
   return (
     <div className={styles.root}>
-      <div className="space-y-2">{children}</div>
+      <div className="flex flex-col items-end space-y-2">{children}</div>
     </div>
   );
 }
