+++ src/components/icons/index.ts
@@ -1,4 +1,7 @@
+export * from './arrow-icon';
 export * from './camera-icon';
+export * from './checkerboard-icon';
 export * from './controller-icon';
 export * from './hashtag-icon';
 export * from './medal-icon';
+export * from './stopwatch-icon';
