+++ src/modules/in-game/components/in-game-current-match.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { ArrowIcon } from '~/components/icons/arrow-icon';
+import { ArrowIcon } from '~/components/icons';
 import * as styles from './in-game-current-match.css';
 
 export interface AOTC2023CurrentMatchProps {
